finance-agent-app/requirements.txt
streamlit
yfinance
transformers
pandas
matplotlib
python-dotenv
textblob
rapidfuzz
scikit-learn
newspaper4k
duckduckgo-search
groq
agno